♻️ 지도 리팩토링: useMapAssets, useMapActiveSync, useMapFilterSync hook으로 분리, 상수 파일로 분리
