+++ src/constants/building.js
@@ -16,3 +16,18 @@ export const SHOW_LOCATION = [
   { value: 'STUDENT_UNION', label: '학생문화관' },
   { value: 'GRASS_GROUND', label: '잔디광장' },
 ];
+
+// 지도 SVG 빌딩 id 목록
+export const BUILDING_IDS = [
+  'STUDENT_UNION',
+  'SPORT_TRACK',
+  'HYUUT_GIL',
+  'HUMAN_ECOLOGY_BUILDING',
+  'HAK_GWAN',
+  'GRASS_GROUND',
+  'EWHA_SHINSEGAE_BUILDING',
+  'EWHA_POSCO_BUILDING',
+  'EDUCATION_BUILDING',
+  'WELCH_RYANG_AUDITORIUM',
+  'SENTENNIAL_MUSEUM',
+];
