+++ src/hooks/index.js
@@ -9,5 +9,8 @@ export { useInfiniteScroll } from './useInfiniteScroll.js';
 export { useInfiniteList, buildQueryParams } from './useInfiniteList.js';
 export { usePopularKeywords } from './usePopularKeywords.js';
 export { useSearchResults } from './useSearchResults.js';
-export { default as useMapFocus } from './useMapFocus.js';
+export { default as useMapFocus } from './map/useMapFocus.js';
+export { default as useMapAssets } from './map/useMapAssets.js';
+export { default as useMapActiveSync } from './map/useMapActiveSync.js';
+export { default as useMapFilterSync } from './map/useMapFilterSync.js';
 export { default as useScrollToTop } from './useScrollToTop.js';
